"relocate controllers and scheme"
