+++ controllers/index.js
@@ -1,0 +1,3 @@
+const controllers = require("./controllers");
+
+module.exports = controllers;
