+++ routes/api/contacts.js
@@ -1,82 +1,35 @@
 const express = require("express");
-const contacts = require("../../models/contacts");
+
 const router = express.Router();
-const { HttpError } = require("../../helpers");
-const Joi = require("joi");
+const controllers = require("../../controllers");
 
-const addScheme = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+router.get(
+  "/",
+  async (req, res, next) => await controllers.getListController(req, res, next)
+);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get(
+  "/:contactId",
+  async (req, res, next) =>
+    await controllers.getContactController(req, res, next)
+);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not Found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post(
+  "/",
+  async (req, res, next) =>
+    await controllers.postContactController(req, res, next)
+);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addScheme.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { name, email, phone } = req.body;
-    const result = await contacts.addContact(name, email, phone);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete(
+  "/:contactId",
+  async (req, res, next) =>
+    await controllers.deleteContactController(req, res, next)
+);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    if (!req.body) {
-      throw HttpError(400, "missing fields");
-    }
-    const { error } = addScheme.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put(
+  "/:contactId",
+  async (req, res, next) =>
+    await controllers.putContactController(req, res, next)
+);
 
 module.exports = router;
