+++ schemes/index.js
@@ -1,0 +1,3 @@
+const addScheme = require("./schemeJoi");
+
+module.exports = addScheme;
